Add files via upload
This script creates a pipeline with PCA and a Decision Tree classifier. It iterates through 11 different subject folders, performs a Grid Search  on parameters, performs Kfold cross-validation. Explained variance and confusion matrices plots are generated and saved to each subject folder and a dataframe of cross-validation scores are saved as CSV data.